Create abstract Game class and TicTacToe implementation
A Game is defined by a representation of states and moves, as well as a
rule for determining if a move is legal in the current state.

TicTacToe is an example implementation of the abstract Game class. We
can use this a model for designing more complex games like chess.